Made Fixes to Front End
I noticed that for Home in the nav bar, it lead to href="/static", which returned a 404 error. I changed this to "/index". I just made this change to my exsisting commit for the nav bar fragment.

I also noticed that for the Taskpage.html, the call for the js scripts were placed after the body and html tags.

Additionally I added "https://" to the href reference of "cdn.datatables.net/2.1.8/css/dataTables.dataTables.min.css". Originally it did not include https:// infront of the link, which I think would not lead to a successful retrival of the js or css.